Fix warnings

+++ chapter01/problem_03/main.cpp
@@ -13,10 +13,10 @@ size_t leastCommonMultiple(const size_t& num1, const size_t& num2) {
 
 // chapter-1 -> problem 3 (least common multiple)
 TEST_CASE("Get least common multiple", "[least_common_multiple]") {
-  REQUIRE(leastCommonMultiple(10, 5) == std::lcm(10, 5));
-  REQUIRE(leastCommonMultiple(15, 3) == std::lcm(15, 3));
-  REQUIRE(leastCommonMultiple(22, 2) == std::lcm(22, 2));
-  REQUIRE(leastCommonMultiple(35, 5) == std::lcm(35, 5));
-  REQUIRE(leastCommonMultiple(50, 5) == std::lcm(50, 5));
-  REQUIRE(leastCommonMultiple(100, 20) == std::lcm(100, 20));
+  REQUIRE(leastCommonMultiple(10, 5) == static_cast<size_t>(std::lcm(10, 5)));
+  REQUIRE(leastCommonMultiple(15, 3) == static_cast<size_t>(std::lcm(15, 3)));
+  REQUIRE(leastCommonMultiple(22, 2) == static_cast<size_t>(std::lcm(22, 2)));
+  REQUIRE(leastCommonMultiple(35, 5) == static_cast<size_t>(std::lcm(35, 5)));
+  REQUIRE(leastCommonMultiple(50, 5) == static_cast<size_t>(std::lcm(50, 5)));
+  REQUIRE(leastCommonMultiple(100, 20) == static_cast<size_t>(std::lcm(100, 20)));
 }
